movie list unit tests have written (parsing mock data and output controls), moviedetail screen mvvm architecture implemented

+++ popularMovieApp/Models/MovieModel.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct MovieModel: Codable {
     var popularity: Double?
-    var voteCount: Double?
+    var voteCount: Int?
     var video: Bool?
     var posterPath: String?
     var id: Int?
